requirements.txt
aniso8601==1.2.0
Flask==1.1.1
cryptography==2.3
pyOpenSSL==17.5.0
PyYAML==3.13
six==1.11.0


Werkzeug~=2.0.2
setuptools~=57.0.0
Jinja2~=3.0.3
asn1crypto~=1.4.0
certvalidator~=0.11.1
python-dateutil~=2.8.2